Fix double-press-to-switch, hopefully.
I'm not sure what *exactly* changed, when, and whether it's the Emacs dbus api
or Ubuntu. It happened while I wasn't looking. But calling dbus language
switch synchronously resulted in losing the language change event most of the
time -- but not *all* of the time.

Calling it asynchronously seems to be the answer.

I'm hoping this fixes #1, but please reopen it if it doesn't work for you.